+++ backend/canisters/community/api/src/updates/mod.rs
@@ -8,6 +8,7 @@ pub mod c2c_bot_delete_channel;
 pub mod c2c_bot_delete_messages;
 pub mod c2c_bot_invite_users;
 pub mod c2c_bot_remove_user;
+pub mod c2c_bot_remove_user_from_channel;
 pub mod c2c_bot_send_message;
 pub mod c2c_bot_subscribe_to_events;
 pub mod c2c_create_proposals_channel;